fix free disk space detection on x32

+++ src/FileSystem/FileSystem.cpp
@@ -737,9 +737,9 @@ unsigned long CFileSystem::getMBsFree(const std::string& path)
 		return 0;
 	}
 	if (st.f_frsize) {
-		return (st.f_frsize * st.f_bavail) / (1024 * 1024);
+		return ((uint64_t)st.f_frsize * st.f_bavail) / (1024 * 1024);
 	}
-	return (st.f_bsize * st.f_bavail) / (1024 * 1024);
+	return ((uint64_t)st.f_bsize * st.f_bavail) / (1024 * 1024);
 #endif
 }
